+++ src/cassandra/schema/table_meta.rs
@@ -79,24 +79,24 @@ impl TableMeta {
 
     /// Gets the total number of columns for the table.
     pub fn column_count(&self) -> usize {
-        unsafe { cass_table_meta_column_count(self.0) }
+        unsafe { cass_table_meta_column_count(self.0) as usize }
     }
 
     /// Gets the column metadata for the provided index.
     pub fn column(&self, index: usize) -> ColumnMeta {
         // TODO: can return NULL
-        unsafe { ColumnMeta::build(cass_table_meta_column(self.0, index)) }
+        unsafe { ColumnMeta::build(cass_table_meta_column(self.0, index as u64)) }
     }
 
     /// Gets the number of columns for the table's partition key.
     pub fn partition_key_count(&self) -> usize {
-        unsafe { cass_table_meta_partition_key_count(self.0) }
+        unsafe { cass_table_meta_partition_key_count(self.0) as usize }
     }
 
     /// Gets the partition key column metadata for the provided index.
     pub fn partition_key(&self, index: usize) -> Option<ColumnMeta> {
         unsafe {
-            let key = cass_table_meta_partition_key(self.0, index);
+            let key = cass_table_meta_partition_key(self.0, index as u64);
             if key.is_null() {
                 None
             } else {
@@ -107,13 +107,13 @@ impl TableMeta {
 
     /// Gets the number of columns for the table's clustering key
     pub fn clustering_key_count(&self) -> usize {
-        unsafe { cass_table_meta_clustering_key_count(self.0) }
+        unsafe { cass_table_meta_clustering_key_count(self.0) as usize }
     }
 
     /// Gets the clustering key column metadata for the provided index.
     pub fn cluster_key(&self, index: usize) -> Option<ColumnMeta> {
         unsafe {
-            let key = cass_table_meta_clustering_key(self.0, index);
+            let key = cass_table_meta_clustering_key(self.0, index as u64);
             if key.is_null() {
                 None
             } else {
